LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.wordSpacing.measure-expected.txt

2d.text.drawing.style.wordSpacing.measure

Testing word spacing with different length units


FAIL Testing word spacing with different length units assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
